Added timeline and unfollow
