+++ backend/instatwitter/instatwitter/urls.py
@@ -18,5 +18,6 @@ from django.urls import path,include
 
 urlpatterns = [
     path('',include('main.urls')),
+    path('user/',include('user_profile.urls')),
     path('admin/', admin.site.urls),
 ]
