+++ backend/instatwitter/main/views.py
@@ -74,7 +74,7 @@ def registration_api(request):
 
 @api_view(['POST'])
 @csrf_exempt
-def username_api(request):
+def home_page_api(request):
     #retrive the token 
     received_token = request.data.get('token',None)
     print(received_token)
@@ -99,60 +99,11 @@ def username_api(request):
             usertweets.append(serializer.data)
         print(usertweets)
 
-        return Response(usertweets,status= status.HTTP_200_OK)
+        return Response([data,usertweets],status= status.HTTP_200_OK)
 
-        return Response(data,status= status.HTTP_200_OK)
+        # return Response(data,status= status.HTTP_200_OK)
     return Response(status= status.HTTP_504_GATEWAY_TIMEOUT)
 
-
-
-@api_view(['POST'])
-@csrf_exempt
-def create_user_profile_api(request):
-    receivedtoken = request.data.get('token',None)
-    print(receivedtoken)
-    
-    if receivedtoken is None:
-         return Response(status= status.HTTP_504_GATEWAY_TIMEOUT)
-    else:
-        profiletoken = Token.objects.get(key= receivedtoken)
-        user = profiletoken.user
-        firstname = user.first_name
-        lastname = user.last_name
-        username = user.username
-        user_bio = UserData.objects.get(user=user)
-       
-        data ={}
-        data['firstname'] = firstname
-        data['lastname'] = lastname
-        data['username'] = username
-        receivedbio = request.data.get('bio',None)
-        receivedfirstname = request.data.get('firstname',None)
-        receivedlastname = request.data.get('lastname',None)
-
-        if receivedbio is not None :
-            if receivedfirstname is not None or receivedlastname is not None:
-                user.first_name = receivedfirstname
-                user.last_name = receivedlastname
-                user.save()
-                data['firstname'] = receivedfirstname
-                data['lastname'] = receivedlastname
-            user_bio.userbio=receivedbio
-            user_bio.save()
-            data['bio'] = user_bio.userbio
-            
-            return Response(data,status= status.HTTP_200_OK)      
-                
-        else:
-            mybio = UserData.objects.get(user= user)
-            data['bio']= mybio.userbio                 
-            return Response(data,status= status.HTTP_200_OK)
-        
-
-        return Response(data)
-    
-   
-    
 
 @api_view(['POST'])
 @csrf_exempt
@@ -204,6 +155,8 @@ def menu_api(request):
         data={}
         #pass the username of that object
         data['username'] = user.username
+        data['firstname'] = user.first_name
+        data['lastname'] = user.last_name
         currentuser = user.username   #set that username name as current 
         #get a list of all the users present, except the logged in one
         myusers = User.objects.values_list('username',flat=True).exclude(username=currentuser)  
@@ -214,136 +167,10 @@ def menu_api(request):
             mytweet.save()      #save the tweet for respective user
 
             #return a list of all users and username of current user along with appropriate response code
-            return Response([data,myusers],status= status.HTTP_200_OK)          
+            return Response([data,myusers],status= status.HTTP_201_CREATED)  
+        # else:
+        #     return Response(status=status.HTTP_204_NO_CONTENT)        
         return Response([data,myusers])     #otherwise send the same data without status code, resulting failure.
     #check if flag is set    
     elif  status_flag is True:
         return Response(status= status.HTTP_200_OK)
-    
-        
-    
-        
-
-
-@api_view(['POST'])
-@csrf_exempt
-def user_profile_api(request):
-    received_token_ = request.data.get('token',None)
-    print(received_token_)
-    if received_token_ is None:
-        return Response(status= status.HTTP_504_GATEWAY_TIMEOUT)
-    else:
-        #using received token to get the user related to it
-        usertoken = Token.objects.get(key=received_token_)
-        user = usertoken.user
-        requireddata={}
-        #retreiving data from the user, to send to UI
-        requireddata['firstname'] = user.first_name
-        requireddata['lastname'] = user.last_name
-        requireddata['username'] = user.username
-        #fetch the bio of the user
-        user_bio = UserData.objects.get(user=user)
-        requireddata['bio'] = user_bio.userbio
-        #sending user's basic information to be set in the UI
-       
-        
-        #Retriving all tweets of a particular user and sending it to be displayed after serializing
-        usertweets = TweetData.objects.filter(user=user).order_by('-time_created')
-        serializer = TweetSerializer(usertweets,many=True)
-        #return the serialized tweets along with the basic information of user with status code       
-        return Response([requireddata,serializer.data], status= status.HTTP_200_OK)
-
-
-@api_view(["POST"])
-@csrf_exempt
-
-def following_api(request):
-    #retrive token from the data
-    receivedToken = request.data.get('token',None)
-    print(receivedToken)
-    
-    #Proceed only if received token is not empty
-    if receivedToken is not None:
-
-        #retrive user related to that token
-        followuser = Token.objects.get(key=receivedToken)
-        currentUser = followuser.user
-
-        #create object of the requested following user 
-        followingobj = Follow.get_following(user=currentUser)
-        output =[]
-        
-        for eachuser in followingobj:
-           #for every user present in following object, retrive it's data
-            serializer = UserSerializer(instance= eachuser)
-            output.append(serializer.data)
-
-       
-        return Response([serializer.data,output],status= status.HTTP_200_OK)
-    return Response(status= status.HTTP_400_BAD_REQUEST)
-
-
-@api_view(['POST'])
-@csrf_exempt
-def follower_api(request):
-    #retrive the given token
-    token = request.data.get('token',None)
-    print(token)
-
-    #Proceed only if received token is not empty
-    if token is not None:
-
-        #retrive user related to that token
-        followertoken = Token.objects.get(key=token)
-        currentuser_ = followertoken.user
-
-        #create onject of the requested follower
-        followerobj = Follow.get_follower(user=currentuser_)
-        output_list=[]
-
-        #for every id number present in follower object; which is a user, retrive it's data
-        for i_d in followerobj:
-            serializer = UserSerializer(instance=i_d)
-            #get only the user name of the user, append it output list and send that list 
-            output_list.append(serializer.data['username'])
-            
-
-        return Response(output_list,status=status.HTTP_200_OK)
-    return Response(status=status.HTTP_400_BAD_REQUEST)
-
-
-
-@api_view(['POST'])
-@csrf_exempt
-
-def view_profile_api(request):
-    #retrive the data from request
-    requsteduser = request.data.get('otherUserName',None)
-    token = request.data.get('token',None)
-   
-
-    #Proceed only if received token is not empty
-    if token is not None:
-
-        #from the received username, get the matching data of that user
-        requsted_user = User.objects.get(username=requsteduser)
-        
-        requesteddata={}
-        requesteddata['firstname']= requsted_user.first_name
-        requesteddata['lastname']= requsted_user.last_name
-        requesteddata['username']= requsted_user.username
-        
-        #retrive bio of that user,depending on its username
-        requstedbio = UserData.objects.get(user=requsted_user.id)
-      
-        requesteddata['bio']= requstedbio.userbio
-        # requested_user_id=(requsted_user.id)
-
-        #retrive fields of that user depending on its username
-        requestedtweets = TweetData.objects.filter(user= requsted_user).order_by('-time_created')
-        alltweets= TweetSerializer(requestedtweets,many=True)
-
-        #send the user information and tweets
-        return Response([requesteddata,alltweets.data],status= status.HTTP_200_OK)
-    return Response(status= status.HTTP_400_BAD_REQUEST)
-
