+++ backend/instatwitter/user_profile/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class UserProfileConfig(AppConfig):
+    name = 'user_profile'
